+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Using geolocation to find closest http server available
+- Decides on torrents based on publish_method remote app property
 
 ### Fixed
 - HTTP downloader resumes the download on error