+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - HTTP downloader resumes the download on error
 - Moving files after installation instead of copying them
+- Fix keys API connection problem
 
 ## [3.1.4]
 ### Fixed