+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Using geolocation to find closest http server available
 - Decides on torrents based on publish_method remote app property
 
+### Changed
+- Display warning message when auto-update fails and app is installed
+
 ### Fixed
 - HTTP downloader resumes the download on error
 - Moving files after installation instead of copying them
+- Fix keys API connection problem
+- Fix API cache usage problems
+- Fix problem with remaining ".app" directories on Mac OSX
 
 ## [3.1.4]
 ### Fixed